Moved commonly used functions to renderers/utils.py
